# Fix row schema of nested queries with ORDER BY clause to get proper row type during planning validation