--- a/sql/src/main/java/org/apache/druid/sql/calcite/rule/DruidRules.java
+++ b/sql/src/main/java/org/apache/druid/sql/calcite/rule/DruidRules.java
@@ -30,11 +30,9 @@ import org.apache.calcite.rel.core.Project;
 import org.apache.calcite.rel.core.Sort;
 import org.apache.druid.java.util.common.StringUtils;
 import org.apache.druid.sql.calcite.planner.PlannerContext;
-import org.apache.druid.sql.calcite.rel.CannotBuildQueryException;
 import org.apache.druid.sql.calcite.rel.DruidOuterQueryRel;
 import org.apache.druid.sql.calcite.rel.DruidRel;
 import org.apache.druid.sql.calcite.rel.PartialDruidQuery;
-import org.apache.druid.utils.Throwables;
 
 import java.util.List;
 import java.util.function.BiFunction;
@@ -225,14 +223,7 @@ public class DruidRules
                              .withSort(sort)
         );
         if (outerQueryRel.isValidDruidQuery()) {
-          try {
-            call.transformTo(outerQueryRel);
-          }
-          catch (Exception ex) {
-            if (Throwables.getCauseOfType(ex, CannotBuildQueryException.class) == null) {
-              throw ex;
-            }
-          }
+          call.transformTo(outerQueryRel);
         }
       }
     };