… sql/src/test/java/org/apache/druid/sql/calcite/CalciteExplainQueryTest.java