Queries like
```
SELECT __time,channel FROM wikipedia
WHERE
(
channel in
(
SELECT  channel FROM wikipedia
))
ORDER BY __time DESC
```
was failing in phase of partial query generation. The change in this PR just makes DruidOuterQueryRel eagerly 'validate' the subquery by explaining it, so that the exception happens early and the 'is valid' check catches and returns false, which allows the planner to proceed to other plans. 

#12418 tried to address an issue of using alias with outer query with an order by clause. The reason was that during the query explaining part the source relation type was being mismatched. In the first run it was set to the name of the column while in the next run it was set to the alias name. To mitigate the difference we set the Data source of the subquery to use a QueryDataSource from a TableDataSource which introduced the defect shown here. 

To solve this we move back to the old implementation and added the check for the rowtype with an additional `toDriodQueryForExplain` call that would ensure setting the correct row types. This additional call does not impact performance as the planning time remained almost similar. A run of our benchmarks for planning indicate that they remain in the same ball park.

**Before this change**
```
Benchmark             (query)  (rowsPerSegment)  (storageType)  (vectorize)  Mode  Cnt    Score   Error  Units
SqlBenchmark.planSql        0           5000000           mmap        force  avgt   10    1.425 ± 1.138  ms/op
SqlBenchmark.planSql        1           5000000           mmap        force  avgt   10    1.094 ± 0.070  ms/op
SqlBenchmark.planSql        2           5000000           mmap        force  avgt   10    1.220 ± 0.300  ms/op
SqlBenchmark.planSql        3           5000000           mmap        force  avgt   10    1.364 ± 0.150  ms/op
SqlBenchmark.planSql        4           5000000           mmap        force  avgt   10    2.393 ± 0.941  ms/op
SqlBenchmark.planSql        5           5000000           mmap        force  avgt   10    1.919 ± 0.198  ms/op
SqlBenchmark.planSql        6           5000000           mmap        force  avgt   10    1.884 ± 0.141  ms/op
SqlBenchmark.planSql        7           5000000           mmap        force  avgt   10    2.197 ± 0.369  ms/op
SqlBenchmark.planSql        8           5000000           mmap        force  avgt   10    3.206 ± 0.356  ms/op
SqlBenchmark.planSql        9           5000000           mmap        force  avgt   10    2.439 ± 0.245  ms/op
SqlBenchmark.planSql       10           5000000           mmap        force  avgt   10    0.981 ± 0.017  ms/op
SqlBenchmark.planSql       11           5000000           mmap        force  avgt   10    1.030 ± 0.049  ms/op
SqlBenchmark.planSql       12           5000000           mmap        force  avgt   10    0.769 ± 0.016  ms/op
SqlBenchmark.planSql       13           5000000           mmap        force  avgt   10    1.185 ± 0.113  ms/op
SqlBenchmark.planSql       14           5000000           mmap        force  avgt   10    1.240 ± 0.053  ms/op
SqlBenchmark.planSql       15           5000000           mmap        force  avgt   10    0.791 ± 0.016  ms/op
SqlBenchmark.planSql       16           5000000           mmap        force  avgt   10    0.963 ± 0.038  ms/op
SqlBenchmark.planSql       17           5000000           mmap        force  avgt   10    1.096 ± 0.008  ms/op
SqlBenchmark.planSql       18           5000000           mmap        force  avgt   10    1.289 ± 0.058  ms/op
SqlBenchmark.planSql       19           5000000           mmap        force  avgt   10  138.326 ± 5.060  ms/op
SqlBenchmark.planSql       20           5000000           mmap        force  avgt   10    0.932 ± 0.029  ms/op
```

**After the change**
```
Benchmark             (query)  (rowsPerSegment)  (storageType)  (vectorize)  Mode  Cnt    Score   Error  Units
SqlBenchmark.planSql        0           5000000           mmap        force  avgt   10    0.727 ± 0.060  ms/op
SqlBenchmark.planSql        1           5000000           mmap        force  avgt   10    0.714 ± 0.023  ms/op
SqlBenchmark.planSql        2           5000000           mmap        force  avgt   10    0.830 ± 0.098  ms/op
SqlBenchmark.planSql        3           5000000           mmap        force  avgt   10    1.007 ± 0.110  ms/op
SqlBenchmark.planSql        4           5000000           mmap        force  avgt   10    1.413 ± 0.219  ms/op
SqlBenchmark.planSql        5           5000000           mmap        force  avgt   10    1.356 ± 0.088  ms/op
SqlBenchmark.planSql        6           5000000           mmap        force  avgt   10    1.804 ± 0.217  ms/op
SqlBenchmark.planSql        7           5000000           mmap        force  avgt   10    1.697 ± 0.251  ms/op
SqlBenchmark.planSql        8           5000000           mmap        force  avgt   10    2.368 ± 0.486  ms/op
SqlBenchmark.planSql        9           5000000           mmap        force  avgt   10    2.430 ± 0.229  ms/op
SqlBenchmark.planSql       10           5000000           mmap        force  avgt   10    1.244 ± 0.404  ms/op
SqlBenchmark.planSql       11           5000000           mmap        force  avgt   10    0.947 ± 0.133  ms/op
SqlBenchmark.planSql       12           5000000           mmap        force  avgt   10    0.746 ± 0.064  ms/op
SqlBenchmark.planSql       13           5000000           mmap        force  avgt   10    1.485 ± 0.176  ms/op
SqlBenchmark.planSql       14           5000000           mmap        force  avgt   10    1.608 ± 0.059  ms/op
SqlBenchmark.planSql       15           5000000           mmap        force  avgt   10    0.965 ± 0.032  ms/op
SqlBenchmark.planSql       16           5000000           mmap        force  avgt   10    1.137 ± 0.072  ms/op
SqlBenchmark.planSql       17           5000000           mmap        force  avgt   10    1.346 ± 0.039  ms/op
SqlBenchmark.planSql       18           5000000           mmap        force  avgt   10    1.539 ± 0.062  ms/op
SqlBenchmark.planSql       19           5000000           mmap        force  avgt   10  154.463 ± 5.760  ms/op
SqlBenchmark.planSql       20           5000000           mmap        force  avgt   10    1.041 ± 0.024  ms/op
```


Separate test cases has been updated for this failed test case and also for #13163 

This PR has:
- [x] been self-reviewed.
   - [ ] using the [concurrency checklist](https://github.com/apache/druid/blob/master/dev/code-review/concurrency.md) (Remove this item if the PR doesn't have any relation to concurrency.)
- [ ] added documentation for new or modified features or behaviors.
- [ ] added Javadocs for most classes and all non-trivial methods. Linked related entities via Javadoc links.
- [ ] added or updated version, license, or notice information in [licenses.yaml](https://github.com/apache/druid/blob/master/dev/license.md)
- [ ] added comments explaining the "why" and the intent of the code wherever would not be obvious for an unfamiliar reader.
- [x] added unit tests or modified existing tests to cover new code paths, ensuring the threshold for [code coverage](https://github.com/apache/druid/blob/master/dev/code-review/code-coverage.md) is met.
- [ ] added integration tests.
- [x] been tested in a test Druid cluster.