@abhishekagarwal87 2022-10-07
@somu-imply - I looked at that benchmark. None of those queries seem to have any complex sub-query. It would be better to hand-craft a couple of nested group-by queries that are going to hit the code path you changed. The nested group-by can have a large in-filter (1000). This is when we know how much time is getting added to planning.