@somu-imply 2022-10-08
Will do so @abhishekagarwal87 . Based on comments from Paul and Gian I'll look into fixing the DruidRel so that the try catch can be avoided