@somu-imply 2022-10-12
@abhishekagarwal87 I have updated the `SqlBenchmarks` with 2 additional cases which hit the new code in `DruidOuterQueryRel`. These cases have inner query with large in filters. I tested the working build (before 2022.04) and the current with the changes.

Before
```
Benchmark             (query)  (rowsPerSegment)  (vectorize)  Mode  Cnt  Score   Error  Units
SqlBenchmark.planSql       21           5000000        false  avgt    5  2.819 ± 0.142  ms/op
SqlBenchmark.planSql       22           5000000        false  avgt    5  2.920 ± 0.306  ms/op
```

After
```
Benchmark             (query)  (rowsPerSegment)  (storageType)  (vectorize)  Mode  Cnt  Score   Error  Units
SqlBenchmark.planSql       21           5000000           mmap        false  avgt    5  3.261 ± 0.299  ms/op
SqlBenchmark.planSql       22           5000000           mmap        false  avgt    5  3.215 ± 0.236  ms/op
```