@somu-imply 2022-10-12
@paul-rogers I have removed the dubious `try-catch` which was trying to patch things as the rowtype should be set correctly before the code reaches there. But doing additional work in `DruidOuterQueryRel` leads to an increase in planning time. The travis passed on the last run and my last commit just had benchmarks updated so I expect travis to pass again